All Task/ConsoleBit/requirement.txt
Django==4.0.4
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0